Updated readme and keymap oled
